refactor github client in a separate project
